requirements.txt
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
python-dateutil==2.9.0.post0
python-telegram-bot==20.3
python-telegram-bot-calendar==1.0.5
requests==2.32.5
six==1.17.0
sniffio==1.3.1
urllib3==2.6.2
